ml-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
scikit-learn==1.4.0
statsmodels==0.14.1
psycopg2-binary==2.9.9
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
joblib==1.3.2
